// src/app/users/models/user-details.ts
export interface UserDetails {
	id: number;
	firstName: string;
	lastName: string;
	email: string;
	phone: string;
	webSite: string;
	avatar: string;
	addressZipCode: string;
	addressCity: string;
	addressStreet: string;
	addressSuit: string;
	companyName: string;
	companyCatchPhrase: string;
	companyBs: string;
}